Introduce CI (#3)
Configures a Rakefile with a default task to run RSpec and standard.

Ran `bundle exec rspec --init` to generate `spec/spec_helper.rb` and
`.rspec`.

Additionally, we add a GitHub Action to run this default task based on
a [starter Ruby workflow][1].

We also ensure that tests are run in the test environment via
`ENV["RACK_ENV"]`

[1]: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-ruby#using-the-ruby-starter-workflow