Build production ready proof of concept
Build the simplest possible application that can be deployed to a
production environment, like Heroku.

Our application simply returns a 200 HTTP status, with no headers, and
renders "Hello World" a plain text response.

We introduce and configure Puma per [Heroku's recommendations][1].
However, we remove the Rails specific configuration. Note how this also
sets the `ENV["RACK_ENV"]` to "development" if not set.

Additionally, we add a [Procfile][2] to run the Puma server as a web
process, and add a binstub to manage our Procfile-based application
locally with [foreman][4], using a [modified version of the
jsbundling-rails][3] binstub.

We also introduce Capybara, and configure our test suite to use our Rack
app.

Finally, we [dynamically manage our gems][5] when the application boots.
We require `bundler` in order for the system `rake` command to work.
However, this is not needed when running `bundle exec rake`.

[1]: https://devcenter.heroku.com/articles/deploying-rails-applications-with-the-puma-web-server#config
[2]: https://devcenter.heroku.com/articles/deploying-rails-applications-with-the-puma-web-server#procfile
[3]: https://github.com/rails/jsbundling-rails/blob/main/lib/install/dev
[4]: https://github.com/ddollar/foreman
[5]: https://bundler.io/guides/groups.html