Create simple homepage
Return an HTML document by leveraging `ERB` to dynamically inject markup
in the body. Note that we now return the correct HTTP Headers declaring
the `Content-Type`.

In order to ensure the body will be rendered correctly, we create a
`render` method that is responsible for setting the `@content` used in
`layout.html.erb` by setting it to the evaluated partial template.

Additionally, we also use this new `render` method to set the headers
and status code. This alleviates us from having to call `render` and
then return the response.