Introduce 404 page
Introduce basic routing so that we can serve a homepage, and handle any
invalid requests with a custom 404 page.

We do this by rendering a static HTML file (similar to Rails) for
increased performance. Because we're introducing a static asset, this is
an opportunity for us to use the [Rack::Static][1] middleware. This
allows us to define where the asset lives, as well as set header rules.
What this means is that the server will respond with a `200` status if
the `/404.html` path is directly requested.

In order to use `Rack::Static`, we introduce [Rack::Builder][2] to
construct our application. If we did not do this, we would have had to
add an additional `when` branch to our `case` statement to handle direct
requests to the `/404.html` path.

[1]: https://www.rubydoc.info/gems/rack/Rack/Static
[2]: https://www.rubydoc.info/gems/rack/Rack/Static