Resolve name servers
Render the name servers to the page after form submission. We continue
to leverage the `@locals` instance variable to store additional data
needed for our partial.

In order to create a simple interface, we rely on the fact that `input
type=url` forces a user to enter a URL. From there, we can leverage
`URI` to get the host, which is needed to resolve the name servers.

Because this test feature creates a network request, we mock the
relevant `Resolv` classes in our test.