Cache response with ETags
Since name servers don't change frequently, we can leverage [ETags][1]
to cache the server's response in the browser. We do this by setting the
values of the name servers to a hash, and storing that value as an
`ETag`. Note that we need to ensure this value is wrapped in double
quotation marks per the specification.

On subsequent requests, the client will set the [If-None-Match][2]
header with that value. allowing us to compare those two values. If they
match, it means the name servers have not changed, and we can serve the
cached request by returning a [304 Not Modified][3] status. We abstract
this logic in the `stale?` method inspired by [Rails][4]

Additionally, we also set the [Cache-Control][5] header to [no-cache][6]
in an effort to ensure we revalidate the response be checking the `ETag`
value.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag
[2]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match
[3]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/304
[4]: https://guides.rubyonrails.org/caching_with_rails.html#conditional-get-support
[5]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control
[6]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control#no-cache