Handle internal server errors
Introduce basic error handling by returning a static HTML file, along
with a `500` status code.

We return a static file to avoid using the application, since the
runtime is already affected by an unhandled exception. Serving a static
file requires less resources.

Additionally, we log the error to the console in an effort to help debug
the issue.