Compress HTTP responses
Introduce `Rack::Deflater` to our middleware. Note that we introduce it
at the top of the stack so that our static assets are also compressed.

This makes it so any client that supports a particular content encoding
(represented by the [Accept-Encoding][1] header in the request) can
understand the compressed request. This is stored in the
[Content-Encoding][2] header.

This is most useful for our large CSS file, but also has an added
benefit of compressing our other requests too.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Encoding
[2]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Encoding