Refactor ETag implementation with built-in middleware
This refactor simplifies our implementation by relying on the existing
[Rack::ConditionalGet][] and [Rack::ETag][] middleware classes.

In order to ensure the ETag is consistently set when resolving name
servers, we need to sort them. This makes it so the response body will
always be the same.

Additionally, this refactor makes it so all requests have a `ETag`.

[Rack::ConditionalGet]: https://github.com/rack/rack/blob/main/lib/rack/conditional_get.rb
[Rack::ETag]: https://github.com/rack/rack/blob/main/lib/rack/etag.rb