Refactor error handling
Use built in middleware to handle exceptions and log to the console.

Introduces `ExceptionHandler` to resurface the error if in
`development` and render it with [Rack::ShowExceptions][]. This is
similar to how Rails renders the web console in development.

If we're not in the `development` environment, we swallow the error and
return our static HTML page. We also use [Rack::Logger][] to create a
better logging mechanism.

[Rack::ShowExceptions]: https://github.com/rack/rack/blob/main/lib/rack/show_exceptions.rb
[Rack::Logger]: https://github.com/rack/rack/blob/main/lib/rack/logger.rb